Move elements management functions to a dedicated file
